Remove `.zipignore`, look for it in the plugin

+++ zip.sh
@@ -2,10 +2,15 @@
 
 plugin_dir="$(dirname "$(dirname "$(realpath "$0")" )" )"
 plugin_slug="$(basename "$plugin_dir")"
-scripts_dir="$plugin_dir/wpps-scripts/"
 
 build_version=`grep 'Version:' $plugin_dir/$plugin_slug.php | cut -f4 -d' '`
 zip_file_name="$plugin_slug.$build_version.zip"
 cd "$(dirname "$plugin_dir")"
-zip -r "$zip_file_name" "$plugin_slug" -x@"$scripts_dir.zipignore"
+
+if [ ! -f "$plugin_dir/.zipignore" ]; then
+	echo "Error: please add a .zipignore to the root of the plugin"
+	exit 1
+fi
+
+zip -r "$zip_file_name" "$plugin_slug" -x@"$plugin_dir/.zipignore"
 mv "$zip_file_name" "$plugin_dir"
